DataIngestion/feature_extraction/parallel/requirements.txt
# Coordinator requirements
redis==5.0.1
psycopg2-binary==2.9.9
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2